+++ src/index.ts
@@ -6,6 +6,7 @@ import importer from "./util/importer";
 import oauthUtil from "./util/oauth.util";
 import {AccessToken} from "./util/oauth.util";
 import assert from "./util/assert";
+import * as types from "./types";
 import { Options, Config, StratoUser, OAuthUser, BlockChainUser, Contract, ContractDefinition, CallArgs, SendTx } from "./types";
 import * as constants from "./constants";
 
@@ -15,6 +16,7 @@ export {
   parser,
   fsUtil,
   importer,
+  types,
   oauthUtil,
   assert,
   StratoUser,